chore: codebase audit — architecture fixes and docs refresh
Fix barrel bypass in enrichment/summarize imports by extracting shared
callout constants. Extract view types to dedicated types.ts. Update all
AGENTS.md, DECISIONS.md, STATUS.md, and ARCHITECTURE.md to reflect
title module and recent features. Two security passes confirm clean.
